fix divers bug,correction esthetiques

+++ commands/owners/gdc.py
@@ -6,6 +6,7 @@
 import interactions,asyncio
 from interactions import Extension,Button,ButtonStyle
 from interactions.ext.prefixed_commands import prefixed_command
+from datetime import datetime
 from utils import console
 
 class Gdc(Extension):
@@ -19,6 +20,8 @@ class Gdc(Extension):
                 bdd=self.bot.bdd
                 status=bdd.get_gamedata("gdc","status")
                 channel=bdd.get_gamedata("gdc","channel")
+                maxmembers=bdd.get_gamedata("gdc","maxmembers")
+                duration=bdd.get_gamedata("gdc","duration")
                 if status == []:
                     bdd.add_gamedata("gdc","status","off")
                     status=bdd.get_gamedata("gdc","status")
@@ -29,14 +32,23 @@ class Gdc(Extension):
                     channel = "Non défini"
                 else:
                     channel = f"<#{channel[0]['datavalue']}>"
+                if maxmembers == []:
+                    bdd.add_gamedata("gdc","maxmembers",3)
+                    maxmembers=bdd.get_gamedata("gdc","maxmembers")
+                if duration == []:
+                    bdd.add_gamedata("gdc","duration",7)
+                    duration=bdd.get_gamedata("gdc","duration")
                 status = status[0]["datavalue"]
-                embed = interactions.Embed(description=f"**Guerre des clans**\n\nStatus : {status} \nSalon : {channel}")
+                embed = interactions.Embed(title="Guerre des clans",description=f"Statut : {status} \nSalon : {channel} \nNombre de membres par clan : {maxmembers[0]['datavalue']} \nDurée : {duration[0]['datavalue']} jours")
                 buttons=[]
                 if status == "on":
                     buttons.append(Button(style=ButtonStyle.DANGER,label="Arrêter",custom_id="gdc_off"))
                 else:
                     buttons.append(Button(style=ButtonStyle.SUCCESS,label="Lancer",custom_id="gdc_on"))
-                buttons.append(Button(style=ButtonStyle.PRIMARY,label="Définir le salon",custom_id="gdc_channel"))
+                # buttons.append(Button(style=ButtonStyle.PRIMARY,label="Gains",custom_id="gdc_gains"))
+                buttons.append(Button(style=ButtonStyle.PRIMARY,label="Salon",custom_id="gdc_channel"))
+                buttons.append(Button(style=ButtonStyle.PRIMARY,label="Membres par clan",custom_id="gdc_members"))
+                buttons.append(Button(style=ButtonStyle.PRIMARY,label="Durée",custom_id="gdc_duration"))
                 if "m" in locals():
                     await m.edit(embed=embed,components=[buttons])
                 else:
@@ -50,9 +62,15 @@ class Gdc(Extension):
                 if i.custom_id == "gdc_on":
                     channel = bdd.get_gamedata("gdc","channel")
                     if channel[0]["datavalue"] == "NO":
-                        await i.send("Vueillez définir le salon de la guerre des clans avant de la lancer !")
+                        await i.send("Veuillez définir le salon de la guerre des clans avant de la lancer !")
                         continue
                     bdd.set_gamedata("gdc","status","on")
+                    start = bdd.get_gamedata("gdc","start")
+                    now = datetime.now().strftime("%Y-%m-%d %H:%M:%S")
+                    if start == []:
+                        bdd.add_gamedata("gdc","start",now)
+                    else:
+                        bdd.set_gamedata("gdc","start",now)
                     await i.send("La guerre des clans a été lancée !")
                     continue
                 if i.custom_id == "gdc_off":
@@ -60,7 +78,7 @@ class Gdc(Extension):
                     await i.send("La guerre des clans a été arrêtée !")
                     continue
                 if i.custom_id == "gdc_channel":
-                    ii=await i.send("Veuillez mentionner le salon de la guerre des clans")
+                    ii=await i.send("Veuillez envoyer le salon de la guerre des clans.")
                     def check(m):
                         return m.message.author.id == ctx.author.id and m.message.channel.id == ctx.channel.id
                     try:
@@ -87,5 +105,43 @@ class Gdc(Extension):
                     bdd.set_gamedata("gdc","channel",data)
                     await ctx.send("Le salon de la guerre des clans a bien été défini !")
                     await channel.delete()
+                    await ii.delete()
+                    continue
+                if i.custom_id == "gdc_members":
+                    ii=await i.send("Veuillez envoyer le nombre de membres par clan.")
+                    def check(m):
+                        return m.message.author.id == ctx.author.id and m.message.channel.id == ctx.channel.id
+                    try:
+                        members = await self.bot.wait_for("message_create",timeout=60,checks=check)
+                        members=members.message
+                    except asyncio.TimeoutError:
+                        continue
+                    try:
+                        data = int(members.content)
+                    except:
+                        await ctx.send("Veuillez entrer un nombre valide !")
+                        continue
+                    bdd.set_gamedata("gdc","maxmembers",data)
+                    await ctx.send("Le nombre de membres par clan a bien été défini !")
+                    await members.delete()
+                    await ii.delete()
+                    continue
+                if i.custom_id == "gdc_duration":
+                    ii=await i.send("Veuillez envoyer la durée de la guerre des clans.")
+                    def check(m):
+                        return m.message.author.id == ctx.author.id and m.message.channel.id == ctx.channel.id
+                    try:
+                        duration = await self.bot.wait_for("message_create",timeout=60,checks=check)
+                        duration=duration.message
+                    except asyncio.TimeoutError:
+                        continue
+                    try:
+                        data = int(duration.content)
+                    except:
+                        await ctx.send("Veuillez entrer un nombre valide !")
+                        continue
+                    bdd.set_gamedata("gdc","duration",data)
+                    await ctx.send("La durée de la guerre des clans a bien été définie !")
+                    await duration.delete()
                     await ii.delete()
                     continue
